pyproject.toml
[tool.poetry]
name = "fastapiproject1"
version = "0.1.0"
description = ""
authors = ["Skelesheit <Skelesheit@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
sqlalchemy = "^2.0.41"
asyncpg = "^0.30.0"
sqlalchemy-utils = "^0.41.2"
alembic = "^1.16.4"
aiosmtplib = "^4.0.1"
email-validator = "^2.2.0"
bcrypt = "^4.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.10.1"
aiofiles = "^24.1.0"
httpx = "^0.28.1"
pydantic-settings = "^2.10.1"
fastapi = "^0.116.1"
redis = "^6.2.0"
celery = "^5.5.3"
typer = "^0.16.0"
pytest = "^8.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
